portfolio_API/requirements.txt
aiosmtplib~=4.0.1
fastapi~=0.115.12
pydantic~=2.11.5
python-dotenv~=1.0.0